Enabled more warnings, switched the deployment target to 10.7 and minor code cleanup

+++ MP42/MP42Ratings.m
@@ -105,7 +105,7 @@
 }
 
 - (NSUInteger) ratingIndexForiTunesCode:(NSString *)aiTunesCode {
-	for (int i = 0; i < [iTunesCodes count]; i++) {
+	for (NSUInteger i = 0; i < [iTunesCodes count]; i++) {
 		if ([[iTunesCodes objectAtIndex:i] isEqualToString:aiTunesCode]) {
 			return i;
 		}
@@ -116,7 +116,7 @@
 - (NSUInteger) ratingIndexForiTunesCountry:(NSString *)aCountry media:(NSString *)aMedia ratingString:(NSString *)aRatingString {
 	NSString *target1 = [[NSString stringWithFormat:@"%@ %@: %@", aCountry, aMedia, aRatingString] lowercaseString];
 	NSString *target2 = [[NSString stringWithFormat:@"%@ %@: %@", aCountry, @"movie & TV", aRatingString] lowercaseString];
-	for (int i = 0; i < [ratings count]; i++) {
+	for (NSUInteger i = 0; i < [ratings count]; i++) {
 		if ([[[ratings objectAtIndex:i] lowercaseString] isEqualToString:target1] || [[[ratings objectAtIndex:i] lowercaseString] isEqualToString:target2]) {
 			return i;
 		}
